refactor identity verification screen to solve commits
