+++ composeApp/src/commonMain/composeResources/values/string.xml
@@ -3,11 +3,11 @@
     <string name="chat">Chat</string>
     <string name="accept_offer">Accept Offer</string>
     <string name="verify_your_identity">Verify Your Identity</string>
-    <string name="optional">(Optional)</string>
+    <string name="optional"> (Optional)</string>
     <string name="verify_identity_description">Uploading your ID helps build trust with customers. Verified craftsmen get more jobs and a special badge on their profile.</string>
     <string name="upload_front_of_national_id">Upload Front of National ID</string>
     <string name="upload_back_of_national_id">Upload Back of National ID</string>
-    <string name="i_ll_verify_later">I'll Verify Later</string>
+    <string name="i_will_verify_later">I'll Verify Later</string>
     <string name="see_nearby_requests">See Nearby Requests</string>
     <string name="tab_to_add_photo">Tap to add photo</string>
 </resources>
